Renaming file E util to a helper file because it doesn't apply to file E anymore
