+++ dataactcore/utils/parentDuns.py
@@ -1,16 +1,15 @@
 import logging
-import pandas as pd
-import time
 import sys
+import time
+
+import pandas as pd
 from sqlalchemy import and_, func
 
+from dataactbroker.helpers.sam_wsdl_helper import config_valid, get_entities
 from dataactbroker.helpers.generic_helper import get_client
-
-from dataactcore.utils.fileE import config_valid, get_entities
 from dataactcore.models.domainModels import DUNS
 from dataactcore.utils.duns import load_duns_by_row
 from dataactcore.models.jobModels import FileType, Submission # noqa
-from dataactcore.models.jobModels import Submission # noqa
 from dataactcore.models.userModel import User # noqa
 
 logger = logging.getLogger(__name__)
